requirements.txt
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dj-database-url==0.5.0
requests==2.27.1
